% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateDataprint.R
\name{generateDataprint}
\alias{generateDataprint}
\title{Generates a Dataprint from a CDM v5 database}
\usage{
generateDataprint(connectionDetails, cdmQualifier, caption, dbms)
}
\arguments{
\item{connectionDetails}{A connectionDetails object created by the DatabaseConnector package.}

\item{cdmQualifier}{The table qualifer for CDM tables (ie, NHANES.dbo)}

\item{dbms}{The SqlRender supported DBMS platform (ie, postgresql, pdw)}
}
\description{
Generates a Dataprint from a CDM v5 database
}
